Split IoT base into new class.
